refactor(ivy): reorganize styling and player files (#26149)
PR Close #26149
