+++ packages/core/src/render3/interfaces/view.ts
@@ -9,7 +9,7 @@
 import {Injector} from '../../di/injector';
 import {QueryList} from '../../linker';
 import {Sanitizer} from '../../sanitization/security';
-import {PlayerHandler} from '../animations/interfaces';
+import {PlayerHandler} from '../interfaces/player';
 
 import {LContainer} from './container';
 import {ComponentQuery, ComponentTemplate, DirectiveDefInternal, DirectiveDefList, PipeDefInternal, PipeDefList} from './definition';
