+++ packages/core/test/bundling/animation_world/bundle.golden_symbols.json
@@ -276,7 +276,7 @@
     "name": "addToViewTree"
   },
   {
-    "name": "allocAnimationContext"
+    "name": "allocPlayerContext"
   },
   {
     "name": "allocStylingContext"
@@ -603,9 +603,6 @@
     "name": "getMultiStartIndex"
   },
   {
-    "name": "getOrCreateAnimationContext"
-  },
-  {
     "name": "getOrCreateInjectable"
   },
   {
@@ -613,6 +610,9 @@
   },
   {
     "name": "getOrCreateNodeInjectorForNode"
+  },
+  {
+    "name": "getOrCreatePlayerContext"
   },
   {
     "name": "getOrCreateRenderer2"
