+++ packages/core/test/render3/styling/core_player_handler_spec.ts
@@ -5,8 +5,8 @@
  * Use of this source code is governed by an MIT-style license that can be
  * found in the LICENSE file at https://angular.io/license
  */
-import {CorePlayerHandler} from '../../../src/render3/animations/core_player_handler';
-import {PlayState} from '../../../src/render3/animations/interfaces';
+import {CorePlayerHandler} from '../../../src/render3/styling/core_player_handler';
+import {PlayState} from '../../../src/render3/interfaces/player';
 import {MockPlayer} from './mock_player';
 
 describe('CorePlayerHandler', () => {
